requirements.txt
uvicorn
fastapi
jaconv
pykakasi
Jinja2
Pillow
reportlab
argparse